Rename MessageBoxTest to MessageBox.Avalonia.Example
