+++ src/MessageBox.Avalonia.Example/MainWindow.xaml
@@ -3,7 +3,7 @@
         xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
         xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
         mc:Ignorable="d" d:DesignWidth="800" d:DesignHeight="450"
-        x:Class="MessageBoxTest.MainWindow"
-        Title="MessageBoxTest">
+        x:Class="MessageBox.Avalonia.Example.MainWindow"
+        Title="MessageBox.Avalonia.Example">
   <Button Name="btnClick" >Click</Button>
 </Window>
