+++ src/MessageBox.Avalonia.Example/Program.cs
@@ -1,9 +1,7 @@
-﻿using System;
-using Avalonia;
-using Avalonia.Controls.ApplicationLifetimes;
+﻿using Avalonia;
 using Avalonia.Logging.Serilog;
 
-namespace MessageBoxTest
+namespace MessageBox.Avalonia.Example
 {
     class Program
     {
