+++ src/MessageBox.Avalonia.Example/nuget.config
